third_party/blink/web_tests/flag-specific/highdpi/inspector-protocol/emulation/device-emulation-partial-expected.txt

Tests that overriding only a single parameter does not affect others.

Running test: collectMetrics

Running test: noOverrides
[PASS]

Running test: width
[PASS]: width=300

Running test: height
[PASS]: height=400

Running test: deviceScaleFactor1
[FAIL]: 1.0000000298023224 instead of 1 for deviceScaleFactor

Running test: deviceScaleFactor2
[FAIL]: 2.0000000596046448 instead of 2 for deviceScaleFactor

Running test: clear
[PASS]

Running test: anotherWidth
[PASS]: width=400

Running test: anotherHeight
[PASS]: height=300

Running test: deviceScaleFactor3
[PASS]: deviceScaleFactor=3

Running test: clear
[PASS]
